teacher dan student fungsi dasarnya sudah bisa

+++ database/migrations/2018_09_29_135110_create_tbl_absensis_table.php
@@ -20,7 +20,7 @@ class CreateTblAbsensisTable extends Migration
             $table->foreign('nisn')->references('nisn')->on('tbl_students');
 
             //kode qr
-            $table->increments('id_qr');
+            $table->unsignedBigInteger('id_qr');
             $table->foreign('id_qr')->references('id_qr')->on('tbl_codes');
 
             $table->timestamps();
